pystocks/requirements.txt
numpy
pandas
yfinance
robin-stocks
python-daemon
psycopg2
sqlalchemy